Is the Phillips Curve Stable and Reliable?
My paper tests the structural break and examines the stability of the Phillips curve by describing the relationship between the inflation rates and the unemployment rates in the United States via using the data of the most recent 31 years based on semiannual data through constructing a simple linear regression model. My paper concludes that the Phillips curve is sensitive to the economic environment and is unstable during the recession period.