HUMAN CAPITAL AND ECONOMIC DEVELOPMENT
Human capital is currently the considerable factor that brings about an economic development, especially the growth of national income or GDP per capita. Many countries in Asia has been trying to raise the quality of education and to achieve a more equitable health-care system to improve human capital. This paper provides a link between improvements in human capital through higher education and health expenditure, and economic growth by generating linear regression model and estimating through ordinary least squares method to see the relationships between observed dependent variables and explanatory variables. The findings obtained from the data across 11 nations in Asia for over the past 15 years (2000 to 2014) has shown that the higher education which is measured by the total enrolment in Tertiary educational level and the health expenditure have significantly positive relationship with GDP per capita, which is in accordance with the theories and expectations. This empirical result has led to the policy implications for which the health care and educational systems should be more supported.